Move get current title to maze utils
