+++ src/dataFetching.ts
@@ -1,6 +1,6 @@
 import { VideoID, getVideo, getVideoID, getYouTubeVideoID } from "../maze-utils/src/video";
 import { ThumbnailSubmission, ThumbnailWithRandomTimeResult } from "./thumbnails/thumbnailData";
-import { getCurrentPageTitle, TitleResult, TitleSubmission } from "./titles/titleData";
+import { TitleResult, TitleSubmission } from "./titles/titleData";
 import { FetchResponse, FetchResponseBinary, sendBinaryRequestToCustomServer } from "../maze-utils/src/background-request-proxy";
 import { BrandingLocation, BrandingResult, CasualVoteInfo, replaceCurrentVideoBranding, updateBrandingForVideo } from "./videoBranding/videoBranding";
 import { logError } from "./utils/logger";
@@ -18,6 +18,7 @@ import { sendRequestToServer } from "./utils/requests";
 import { thumbnailDataCache } from "./thumbnails/thumbnailDataCache";
 import { getAutoWarning } from "./submission/autoWarning";
 import { fetchVideoMetadata, isLiveSync } from "../maze-utils/src/metadataFetcher";
+import { getCurrentPageTitle } from "../maze-utils/src/elements";
 
 interface VideoBrandingCacheRecord extends BrandingResult {
     lastUsed: number;
