+++ src/submission/SubmissionComponent.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { CustomThumbnailResult, ThumbnailSubmission } from "../thumbnails/thumbnailData";
-import { getCurrentPageTitle, TitleSubmission } from "../titles/titleData";
+import { TitleSubmission } from "../titles/titleData";
 import { BrandingResult } from "../videoBranding/videoBranding";
 import { ThumbnailType } from "./ThumbnailComponent";
 import { RenderedThumbnailSubmission, ThumbnailDrawerComponent } from "./ThumbnailDrawerComponent";
@@ -27,6 +27,7 @@ import { FormattedText } from "../popup/FormattedTextComponent";
 import { isAutoWarningShown } from "./autoWarning";
 import { getAntiTranslatedTitle } from "../titles/titleAntiTranslateData";
 import { isLiveSync } from "../../maze-utils/src/metadataFetcher";
+import { getCurrentPageTitle } from "../../maze-utils/src/elements";
 
 export interface SubmissionComponentProps {
     videoID: VideoID;
