+++ src/titles/titleData.ts
@@ -1,5 +1,4 @@
 import { BrandingUUID } from "../videoBranding/videoBranding";
-import { getYouTubeTitleNode } from "../../maze-utils/src/elements"
 
 export interface TitleSubmission {
     title: string;
@@ -10,24 +9,4 @@ export interface TitleResult extends TitleSubmission {
     votes: number;
     locked: boolean;
     UUID: BrandingUUID;
-}
-
-export function getCurrentPageTitle(): string | null {
-    const titleNode = getYouTubeTitleNode();
-
-    if (titleNode) {
-        const formattedText = titleNode.querySelector("yt-formatted-string.ytd-watch-metadata, .slim-video-information-title .yt-core-attributed-string:not(.cbCustomTitle)") as HTMLElement;
-        if (formattedText) {
-            return formattedText.innerText;
-        } else {
-            for (const elem of titleNode.children) {
-                if (elem.nodeName === "#text" && elem.nodeValue 
-                        && elem.nodeValue.trim() !== "") {
-                    return elem.nodeValue;
-                }
-            }
-        }
-    }
-
-    return null;
 }
